OHRM5X-2488: refactor
* Renamed Installer\Util\PasswordStrengthService.php to InstallerPasswordStrengthService.php
* Extended from PasswordStrengthService.php
* Changed functions from private to protected
