+++ src/plugins/orangehrmAuthenticationPlugin/Service/PasswordStrengthService.php
@@ -75,7 +75,7 @@ class PasswordStrengthService
     /**
      * @return int
      */
-    private function getMinLength(): int
+    protected function getMinLength(): int
     {
         return $this->minLength ??= $this->getConfigService()->getConfigDao()->getValue(
             ConfigService::KEY_MIN_PASSWORD_LENGTH
@@ -85,7 +85,7 @@ class PasswordStrengthService
     /**
      * @return int
      */
-    private function getMaxLength(): int
+    protected function getMaxLength(): int
     {
         return $this->maxLength ??= ConfigService::MAX_PASSWORD_LENGTH;
     }
@@ -93,7 +93,7 @@ class PasswordStrengthService
     /**
      * @return int
      */
-    private function getMinNoOfLowercaseLetters(): int
+    protected function getMinNoOfLowercaseLetters(): int
     {
         return $this->minNoOfLowercaseLetters ??= $this->getConfigService()->getConfigDao()->getValue(
             ConfigService::KEY_MIN_LOWERCASE_LETTERS
@@ -103,7 +103,7 @@ class PasswordStrengthService
     /**
      * @return int
      */
-    private function getMinNoOfUppercaseLetters(): int
+    protected function getMinNoOfUppercaseLetters(): int
     {
         return $this->minNoOfUppercaseLetters ??= $this->getConfigService()->getConfigDao()->getValue(
             ConfigService::KEY_MIN_UPPERCASE_LETTERS
@@ -113,7 +113,7 @@ class PasswordStrengthService
     /**
      * @return int
      */
-    private function getMinNoOfNumbers(): int
+    protected function getMinNoOfNumbers(): int
     {
         return $this->minNoOfNumbers ??= $this->getConfigService()->getConfigDao()->getValue(
             ConfigService::KEY_MIN_NUMBERS_IN_PASSWORD
@@ -123,7 +123,7 @@ class PasswordStrengthService
     /**
      * @return int
      */
-    private function getMinNoOfSpecialCharacters(): int
+    protected function getMinNoOfSpecialCharacters(): int
     {
         return $this->minNoOfSpecialCharacters ??= $this->getConfigService()->getConfigDao()->getValue(
             ConfigService::KEY_MIN_SPECIAL_CHARACTERS
@@ -133,7 +133,7 @@ class PasswordStrengthService
     /**
      * @return string
      */
-    private function getIsSpacesAllowed(): string
+    protected function getIsSpacesAllowed(): string
     {
         return $this->isSpacesAllowed ??= $this->getConfigService()->getConfigDao()->getValue(
             ConfigService::KEY_IS_SPACES_ALLOWED
@@ -143,7 +143,7 @@ class PasswordStrengthService
     /**
      * @return string
      */
-    private function getDefaultPasswordStrength(): string
+    protected function getDefaultPasswordStrength(): string
     {
         return $this->defaultPasswordStrength ??= $this->getConfigService()->getConfigDao()->getValue(
             ConfigService::KEY_DEFAULT_PASSWORD_STRENGTH
@@ -154,7 +154,7 @@ class PasswordStrengthService
      * @param string $password
      * @return bool
      */
-    private function checkMinPasswordLength(string $password): bool
+    protected function checkMinPasswordLength(string $password): bool
     {
         $minLength = $this->getMinLength();
 
@@ -168,7 +168,7 @@ class PasswordStrengthService
      * @param string $password
      * @return bool
      */
-    private function checkMaxPasswordLength(string $password): bool
+    protected function checkMaxPasswordLength(string $password): bool
     {
         $maxLength = $this->getMaxLength();
 
@@ -183,7 +183,7 @@ class PasswordStrengthService
      * @param string $password
      * @return bool
      */
-    private function checkMinLowercaseLetters(string $password): bool
+    protected function checkMinLowercaseLetters(string $password): bool
     {
         $minNoOfLowercaseLetters = $this->getMinNoOfLowercaseLetters();
 
@@ -198,7 +198,7 @@ class PasswordStrengthService
      * @param string $password
      * @return bool
      */
-    private function checkMinUppercaseLetters(string $password): bool
+    protected function checkMinUppercaseLetters(string $password): bool
     {
         $minNoOfUppercaseLetters = $this->getMinNoOfUppercaseLetters();
         $noOfUppercaseLetters = preg_match_all(self::UPPERCASE_REGEX, $password);
@@ -212,7 +212,7 @@ class PasswordStrengthService
      * @param string $password
      * @return bool
      */
-    private function checkMinNumbersInPassword(string $password): bool
+    protected function checkMinNumbersInPassword(string $password): bool
     {
         $minNoOfNumbers = $this->getMinNoOfNumbers();
         $noOfNumbers = preg_match_all(self::NUMBER_REGEX, $password);
@@ -226,7 +226,7 @@ class PasswordStrengthService
      * @param string $password
      * @return bool
      */
-    private function checkMinSpecialCharacters(string $password): bool
+    protected function checkMinSpecialCharacters(string $password): bool
     {
         $minNoOfSpecialCharacters = $this->getMinNoOfSpecialCharacters();
         $noOfSpecialCharacters = preg_match_all(self::SPECIAL_CHARACTER_REGEX, $password);
@@ -240,7 +240,7 @@ class PasswordStrengthService
      * @param string $password
      * @return bool
      */
-    private function checkSpacesInPassword(string $password): bool
+    protected function checkSpacesInPassword(string $password): bool
     {
         $isSpacesAllowed = $this->getIsSpacesAllowed();
         return $isSpacesAllowed === 'false' && preg_match_all(self::SPACES_REGEX, $password) > 0;
@@ -250,7 +250,7 @@ class PasswordStrengthService
      * @param int $passwordStrength
      * @return bool
      */
-    private function checkRequiredDefaultPasswordStrength(int $passwordStrength): bool
+    protected function checkRequiredDefaultPasswordStrength(int $passwordStrength): bool
     {
         $defaultPasswordStrength = $this->getDefaultPasswordStrength();
         return (($defaultPasswordStrength === 'veryWeak' && $passwordStrength < 0)
